feat(compat): Adds a futures-stream compatibility module
Adds compatibility support for futures::Stream and moves the LegacyStream implementation to a separate module.
Removes unnecessary futures dependencies from the main monoio crate.
