+++ monoio-compat/Cargo.toml
@@ -11,12 +11,15 @@ repository = "https://github.com/bytedance/monoio"
 version = "0.2.2"
 
 [dependencies]
-monoio = { version = "0.2.3", path = "../monoio", default-features = false, features = ["legacy"] }
+monoio = { version = "0.2.3", path = "../monoio", default-features = false, features = [
+    "legacy",
+] }
 reusable-box-future = "0.2"
 tokio = { version = "1", default-features = false, features = ["io-util"] }
 
 hyper = { version = "1.1", optional = true }
 pin-project-lite = { version = "0.2", optional = true }
+futures = { version = "0.3", optional = true }
 
 [dev-dependencies]
 monoio = { version = "0.2.3", path = "../monoio", features = [
@@ -29,3 +32,5 @@ monoio = { version = "0.2.3", path = "../monoio", features = [
 unstable = ["monoio/unstable"]
 # this is an experimental feature
 hyper = ["dep:hyper", "dep:pin-project-lite", "monoio/poll-io"]
+# futures stream compat
+futures-stream = ["dep:futures"]
